+++ sqlrunner/utils.js
@@ -1,0 +1,5 @@
+function unpackObject(object, properties) {
+    return properties.map(property => object[property]);
+}
+
+module.exports = {unpackObject};